feat: 이미지 텍스트 추출에 따른 Alarm과 Calendar 객체와 기능을 수정한다 (#28)
* feat: 처방전 추출 구현

* feat: 처방전 인식을 하여 mediInfo 생성

* feat: 이미지텍스트 추출에 따른 Alarm 도메인의 MVC 재구성

* feat: 과거 test오류로 인한 주석처리

* feat: 매일 복용상태 확인을위한 알람 수정

* feat: alarm 객체와 calendar 객체 MVC 수정

---------

+++ src/main/java/com/kgu/bravoHealthPark/domain/alarm/controller/AlarmController.java
@@ -2,26 +2,28 @@ package com.kgu.bravoHealthPark.domain.alarm.controller;
 
 import com.kgu.bravoHealthPark.domain.alarm.domain.Alarm;
 import com.kgu.bravoHealthPark.domain.alarm.domain.AlarmStatus;
+import com.kgu.bravoHealthPark.domain.alarm.domain.Meal;
 import com.kgu.bravoHealthPark.domain.alarm.dto.AlarmDto;
 import com.kgu.bravoHealthPark.domain.alarm.dto.AlarmForm;
 import com.kgu.bravoHealthPark.domain.alarm.service.AlarmService;
 import com.kgu.bravoHealthPark.domain.medicationInfo.domain.MedicationInfo;
 import com.kgu.bravoHealthPark.domain.medicationInfo.service.MedicationInfoService;
-import com.kgu.bravoHealthPark.domain.user.domain.User;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
+import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Api(tags = {"Alarm Api"}, description = "알람 관련 Api (#8)")
+@Slf4j
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/alarm")
@@ -30,34 +32,55 @@ public class AlarmController {
     private final MedicationInfoService medicationInfoService;
 
     @ApiOperation(value = "알람 생성")
-    @PostMapping("/create")
-    public ResponseEntity<AlarmDto> createAlarm(Long medicationInfoId, String title, String time) {
+    @PostMapping("/{medicationInfoId}")
+    public ResponseEntity<List<AlarmDto>> createAlarm(@PathVariable Long medicationInfoId, Meal meal, String... times) {
         MedicationInfo medicationInfo = medicationInfoService.findByMedicationInfoId(medicationInfoId);
+        LocalDate startDate = medicationInfo.getStartDate();
+        LocalDate endDate = medicationInfo.getEndDate();
+        List<AlarmDto> alarmDtoList = new ArrayList<>();
+
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("HH:mm");
-        LocalTime localTime = LocalTime.parse(time, formatter);
 
-        Alarm alarm = new Alarm(medicationInfo, title, localTime);
-        alarm.initStatus();
+        for (LocalDate currentDate = startDate; !currentDate.isAfter(endDate); currentDate = currentDate.plusDays(1)) {
+            for (String time : times) {
+                LocalTime localTime = LocalTime.parse(time, formatter);
 
-        alarmService.save(alarm);
+                if (meal == Meal.BEFORE_MEAL) {
+                    localTime = localTime.minusMinutes(30);
+                } else {
+                    localTime = localTime.plusMinutes(30);
+                }
 
-        AlarmDto alarmDto = new AlarmDto(alarm);
+                Alarm alarm = new Alarm(medicationInfo, "약 먹을 시간입니다", localTime, meal, currentDate);
+                alarm.initStatus();
+                alarmService.save(alarm);
 
-        return ResponseEntity.ok().body(alarmDto);
+                AlarmDto alarmDto = new AlarmDto(alarm);
+                alarmDtoList.add(alarmDto);
+            }
+        }
+
+        return ResponseEntity.ok().body(alarmDtoList);
     }
 
     @ApiOperation(value = "알람 삭제")
-    @DeleteMapping("/delete/{alarmId}")
+    @DeleteMapping("/{alarmId}")
     public ResponseEntity<AlarmDto> deleteAlarm(@PathVariable Long alarmId) {
         Alarm findAlarm = alarmService.findAlarmById(alarmId);
-        alarmService.deleteAlarm(findAlarm);
+
+        // 연관된 객체를 해제
+        MedicationInfo medicationInfo = findAlarm.getMedicationInfo();
+        if (medicationInfo != null) {
+            findAlarm.deleteMedicationInfo();
+            alarmService.deleteAlarm(findAlarm);
+        }
 
         return ResponseEntity.ok().body(null);
     }
 
     @ApiOperation(value = "알람 수정")
     @PatchMapping("/update/{alarmId}")
-    public ResponseEntity<AlarmDto> updateAlarm(@PathVariable Long alarmId, @RequestParam AlarmForm form) {
+    public ResponseEntity<AlarmDto> updateAlarm(@PathVariable Long alarmId, @RequestBody AlarmForm form) {
         Alarm findAlarm = alarmService.findAlarmById(alarmId);
         alarmService.updateAlarm(findAlarm, form);
 
@@ -66,8 +89,8 @@ public class AlarmController {
         return ResponseEntity.ok().body(alarmDto);
     }
 
-    @ApiOperation(value = "알람 확인후 복용 상태 ")
-    @PatchMapping("/update/alarmstatus/dose/{alarmId}")
+    @ApiOperation(value = "알람 확인후 복용 상태 변경")
+    @PatchMapping("/update/dose/{alarmId}")
     public ResponseEntity<AlarmDto> updateAlarmStatusDose(@PathVariable Long alarmId) {
         Alarm findAlarm = alarmService.findAlarmById(alarmId);
         alarmService.changeAlarmDose(findAlarm);
@@ -77,8 +100,8 @@ public class AlarmController {
         return ResponseEntity.ok().body(alarmDto);
     }
 
-    @ApiOperation(value = "알람 확인후 복용하지 않음 상태")
-    @PatchMapping("/update/alarmstatus/notdose/{alarmId}")
+    @ApiOperation(value = "알람 확인후 복용하지 않음 상태로 변경")
+    @PatchMapping("/update/notdose/{alarmId}")
     public ResponseEntity<AlarmDto> updateUncheck(@PathVariable Long alarmId) {
         Alarm findAlarm = alarmService.findAlarmById(alarmId);
         alarmService.changeAlarmNotDose(findAlarm);
@@ -89,7 +112,7 @@ public class AlarmController {
     }
 
     @ApiOperation(value = "알람 Id로 찾기")
-    @GetMapping("/search/alarm/{alarmId}")
+    @GetMapping("/{alarmId}")
     public ResponseEntity<AlarmDto> searchAlarmById(@PathVariable Long alarmId) {
         Alarm findAlarm = alarmService.findAlarmById(alarmId);
 
@@ -98,22 +121,14 @@ public class AlarmController {
         return ResponseEntity.ok().body(result);
     }
 
-    @ApiOperation(value = "알람 전체 찾기")
-    @GetMapping("/all")
-    public ResponseEntity<List<AlarmDto>> searchAllAlarm() {
-        List<Alarm> alarmList = alarmService.findAlarmAll();
+    @ApiOperation(value = "유저별 알람 찾기")
+    @GetMapping("/{userId}")
+    public ResponseEntity<List<AlarmDto>> searchAlarmByUser(@PathVariable Long userId) {
+        List<Alarm> alarmList = alarmService.findAlarmByUser(userId);
 
-        List<AlarmDto> result = alarmList.stream()
-                .map(a -> new AlarmDto(a))
-                .collect(Collectors.toList());
-
-        return ResponseEntity.ok().body(result);
-    }
-
-    @ApiOperation(value = "알람 이름으로 알람 찾기")
-    @GetMapping("/search/title")
-    public ResponseEntity<List<AlarmDto>> searchAlarmByTitle(@RequestParam String title) {
-        List<Alarm> alarmList = alarmService.findAlarmByTitle(title);
+        if (alarmList.isEmpty()) {
+            log.debug("정보가 없습니다.");
+        }
 
         List<AlarmDto> result = alarmList.stream()
                 .map(a -> new AlarmDto(a))
@@ -123,22 +138,13 @@ public class AlarmController {
     }
 
     @ApiOperation(value = "알람 상태로 알람 찾기")
-    @GetMapping("/search/status")
-    public ResponseEntity<List<AlarmDto>> searchAlarmByStatus(@RequestParam AlarmStatus alarmStatus) {
+    @GetMapping("/status")
+    public ResponseEntity<List<AlarmDto>> searchAlarmByStatus(AlarmStatus alarmStatus) {
         List<Alarm> alarmList = alarmService.findAlarmByStatus(alarmStatus);
 
-        List<AlarmDto> result = alarmList.stream()
-                .map(a -> new AlarmDto(a))
-                .collect(Collectors.toList());
-
-        return ResponseEntity.ok().body(result);
-    }
-
-    @ApiOperation(value = "유저별 알람 찾기")
-    @GetMapping("/search/user/{userId}")
-    public ResponseEntity<List<AlarmDto>> searchAlarmByUser(@PathVariable Long userId) {
-        List<Alarm> alarmList = alarmService.findAlarmByUser(userId);
-
+        if (alarmList.isEmpty()) {
+            log.debug("정보가 없습니다.");
+        }
         List<AlarmDto> result = alarmList.stream()
                 .map(a -> new AlarmDto(a))
                 .collect(Collectors.toList());
@@ -147,9 +153,13 @@ public class AlarmController {
     }
 
     @ApiOperation(value = "유저별 알람 상태에 따라알람 찾기")
-    @GetMapping("/search/user/status/{userId}")
-    public ResponseEntity<List<AlarmDto>> searchAlarmByUserAndStatus(@PathVariable Long userId, @RequestParam AlarmStatus alarmStatus) {
+    @GetMapping("/status/{userId}")
+    public ResponseEntity<List<AlarmDto>> searchAlarmByUserAndStatus(@PathVariable Long userId, AlarmStatus alarmStatus) {
         List<Alarm> alarmList = alarmService.findAlarmByUserAndStatus(userId, alarmStatus);
+
+        if (alarmList.isEmpty()) {
+            log.debug("정보가 없습니다.");
+        }
 
         List<AlarmDto> result = alarmList.stream()
                 .map(a -> new AlarmDto(a))
