+++ src/main/java/com/kgu/bravoHealthPark/domain/calendar/controller/CalendarController.java
@@ -12,6 +12,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -25,29 +26,39 @@ public class CalendarController {
     private final CalendarService calendarService;
 
     @ApiOperation(value = "캘린더 생성")
-    @PostMapping("/create")
-    public ResponseEntity<CalendarDto> createCalendar(Long alarmId, int year, int month, int day) {
-        Alarm alarm = alarmService.findAlarmById(alarmId);
-        Calendar calendar = new Calendar(alarm, alarm.getAlarmStatus(), year, month, day);
+    @PostMapping("/{userId}")
+    public ResponseEntity<List<CalendarDto>> createCalendar(@PathVariable Long userId) {
+        List<Alarm> alarmList = alarmService.findAlarmByUser(userId);
 
-        calendarService.save(calendar);
-        CalendarDto calendarDto = new CalendarDto(calendar);
+        ArrayList<CalendarDto> calendarDtos = new ArrayList<>();
 
-        return ResponseEntity.ok().body(calendarDto);
+        for (Alarm alarm : alarmList) {
+            Calendar calendar = new Calendar(alarm, alarm.getAlarmStatus(),
+                    alarm.getDate().getYear(), alarm.getDate().getMonthValue(), alarm.getDate().getDayOfMonth());
+
+            calendarService.save(calendar);
+            CalendarDto calendarDto = new CalendarDto(calendar);
+            calendarDtos.add(calendarDto);
+        }
+
+        return ResponseEntity.ok().body(calendarDtos);
     }
 
     @ApiOperation(value = "캘린더 삭제")
-    @DeleteMapping("/create")
-    public ResponseEntity createCalendar(@PathVariable Long calendarId) {
+    @DeleteMapping("/{calendarId}")
+    public ResponseEntity<?> deleteCalendar(@PathVariable Long calendarId) {
         Calendar findCalendar = calendarService.findCalendarById(calendarId);
-        calendarService.delete(findCalendar);
-
+        Alarm alarm = findCalendar.getAlarm();
+        if (alarm != null) {
+            findCalendar.deleteAlarm();
+            calendarService.delete(findCalendar);
+        }
         return ResponseEntity.ok().body(null);
     }
 
     @ApiOperation(value = "캘린더 Id로 찾기")
-    @GetMapping("/search/{calendarId}")
-    public ResponseEntity<CalendarDto> searchCalendarById(@PathVariable Long calendarId) {
+    @GetMapping("")
+    public ResponseEntity<CalendarDto> searchCalendarById(@RequestParam Long calendarId) {
         Calendar calendar = calendarService.findCalendarById(calendarId);
 
         CalendarDto calendarDto = new CalendarDto(calendar);
@@ -56,9 +67,9 @@ public class CalendarController {
     }
 
     @ApiOperation(value = "캘린더 전체 찾기")
-    @GetMapping("/search/all")
-    public ResponseEntity<List<CalendarDto>> searchCalendarAll() {
-        List<Calendar> calendarList = calendarService.findCalendarAll();
+    @GetMapping("/all")
+    public ResponseEntity<List<CalendarDto>> searchCalendarAll(@RequestParam Long userId) {
+        List<Calendar> calendarList = calendarService.findCalendarAllByUser(userId);
 
         List<CalendarDto> result = calendarList.stream()
                 .map(c -> new CalendarDto(c))
@@ -68,7 +79,7 @@ public class CalendarController {
     }
 
     @ApiOperation(value = "복약 상태로 찾기")
-    @GetMapping("/search/status")
+    @GetMapping("/status")
     public ResponseEntity<List<CalendarDto>> searchCalendar(@RequestParam AlarmStatus alarmStatus) {
         List<Calendar> calendarByStatus = calendarService.findCalendarByStatus(alarmStatus);
